requirements.txt
numpy==2.0.1
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.3
nltk

